OBPIH-5805 Fix after review
